requirements.txt
altgraph==0.17.4
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mega.py==1.0.8
packaging==23.2
pefile==2023.2.7
pycryptodome==3.20.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
pyinstaller-versionfile==2.1.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tenacity==5.1.5
urllib3==2.2.1
